news-service/requirements.txt
fastapi
uvicorn[standard]
apscheduler
supabase
httpx
beautifulsoup4
lxml
feedparser
python-dotenv
pydantic
transformers
torch
tiktoken
sentencepiece
protobuf
yfinance
vnstock>=3.4.0
opencv-python